refactor(runtime): relocate stage-engine seams into leviath-runtime
Move the decoupling seams and pure engine helpers that the relocated stage
engine will depend on out of leviath-cli into leviath-runtime, so a later
phase can move the stage loop itself without a runtime -> cli cycle:

- RunIO trait -> runtime::run_io (ConsoleIO/MockIO stay in cli)
- StageToolSource + ToolCaller -> runtime::tool_source (RegistryToolCaller
  + impl for ToolRegistry stay in cli)
- InteractionBackend trait + dispatch/gate helpers -> runtime::dynamic_interaction
  (concrete foreground/worker backends stay in cli)
- graph traversal (graph.rs) -> runtime::graph
- ProviderCreds + build_provider_registry(&[ProviderCreds]) ->
  runtime::provider_creds (Config-based translators stay in cli)
- spawn_child_agent -> runtime::spawn
- initialize_context_window + swap_context_layout -> runtime::context_setup

cli re-exports each under its historical commands::run::<name> / crate::tools
path so all existing call sites keep resolving. Behavior unchanged; workspace
builds, clippy (warnings=errors), all tests, and check-exclusions pass.
`cargo tree -p leviath-runtime` has no leviath-cli edge.
