+++ crates/leviath-cli/src/commands/run/tool_source.rs
@@ -1,37 +1,7 @@
-//! Decoupling seam between the stage loop and the concrete tool registry.
+//! Re-export of the tool-source seam, which now lives in `leviath-runtime`.
 //!
-//! Phase 3 of the engine relocation. The stage loop / stages / fan-out only
-//! need two things from the CLI's [`ToolRegistry`](crate::tools::ToolRegistry):
-//! the set of tool definitions to advertise, and a way to execute a single tool
-//! call by name. Capturing exactly that in [`StageToolSource`] lets a later
-//! phase move the stage loop into `leviath-runtime` without a `runtime -> cli`
-//! dependency on `ToolRegistry`. The concrete `impl`s stay in `tools.rs`.
+//! The concrete `RegistryToolCaller` + `impl StageToolSource for ToolRegistry`
+//! stay in the CLI (`crate::tools`); only the traits moved into the runtime so
+//! the relocated stage engine can name them without a `runtime -> cli` cycle.
 
-use std::sync::Arc;
-
-use async_trait::async_trait;
-use leviath_providers::Tool;
-
-/// Executes a single tool call by name.
-///
-/// Kept separate from [`StageToolSource`] because fan-out workers run in
-/// detached (`'static`) tasks and therefore need an *owned*, cheap-to-clone,
-/// `Send + Sync` executor rather than a borrow of the registry.
-#[async_trait]
-pub trait ToolCaller: Send + Sync {
-    /// Execute `name` with `arguments`, returning the tool's textual result
-    /// (errors are rendered into the string, matching `ToolRegistry::call`).
-    async fn call(&self, name: &str, arguments: serde_json::Value) -> String;
-}
-
-/// What the stage loop / stages / fan-out need from the tool registry.
-pub trait StageToolSource: Send + Sync {
-    /// All tool definitions to advertise to the model (built-ins + MCP +
-    /// sub-agent tools). The stage loop and fan-out do their own per-stage
-    /// `available_tools` filtering on top of this.
-    fn all_tool_defs(&self) -> Vec<Tool>;
-
-    /// An owned, `'static` handle for executing tool calls, used by fan-out
-    /// workers spawned as detached tasks.
-    fn tool_caller(&self) -> Arc<dyn ToolCaller>;
-}
+pub use leviath_runtime::tool_source::{StageToolSource, ToolCaller};
