+++ crates/leviath-runtime/src/lib.rs
@@ -7,12 +7,19 @@
 //! are entities and their behaviors are systems.
 
 pub mod components;
+pub mod context_setup;
+pub mod dynamic_interaction;
 pub mod engine;
+pub mod graph;
 pub mod pool;
+pub mod provider_creds;
 pub mod repetition;
+pub mod run_io;
 pub mod scheduler;
+pub mod spawn;
 pub mod systems;
 pub mod taint;
+pub mod tool_source;
 // test_support.rs gates itself with an inner `#![cfg(test)]` attribute, so no
 // `#[cfg(test)]` is needed here (adding one would trigger clippy's
 // `duplicated_attributes` lint under `-D warnings`).
@@ -28,5 +35,8 @@ pub use engine::{
     ToolResultsFuture,
 };
 pub use pool::AgentPool;
+pub use provider_creds::{build_provider_registry, ProviderCreds};
+pub use run_io::RunIO;
 pub use scheduler::TaskScheduler;
+pub use spawn::spawn_child_agent;
 pub use taint::TaintGate;
